base: Introduce Accidental Touch
Accidental Touch is a feature useful for many, especially the gamers out there.
It prevents any sort of accidental touch on hardware buttons while the touchscreen is
in use.
This means you won’t experience any unexpected behaviour while playing games,
web browsing or even when attempting to reach your finger out to the edge of
that humongous display.

Signed-off-by: DennySPb <dennyspb@gmail.com>
Change-Id: I2fe149de65da357b6d37ef1192e4d8fa5375e91d
Signed-off-by: Dmitrii <bankersenator@gmail.com>