Fix Google dialer FC due to missing permissions
In past verisons you could not download the Google dialer unless
you were on a Nexus/Pixel. Now that we're able to download it on pretty
much every device running 8.1, we need to protect the user from the user.

Some users are smarter than others, is that simple. Some users will realize
that the reason this FCs is because of missing permissions. Some won't be that
smart and will come to us and report this as a bug. This commit fixes this.

[AgentFabulous - POSP]
- Adapt for android-10.0.0_r2
- Use grantSystemFixedPermissionsToSystemPackage to simplify syntax

Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: DennySPb <dennyspb@gmail.com>
Change-Id: Id56d5fddcf9088d6b7b3ac205c9a8161fe7b8f9e
Signed-off-by: Dmitrii <bankersenator@gmail.com>